requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
python-dotenv
pyyaml
langchain
langchain-openai
langchain-text-splitters
tiktoken